fix: add static_cast

+++ src/core/ModelImpl.cpp
@@ -215,7 +215,7 @@ void load_indexes(const fastgltf::Asset& gltf, std::vector<uint32_t>& indices,
 
         fastgltf::iterateAccessor<std::uint32_t>(
                 gltf, indexaccessor, [&](std::uint32_t idx) {
-                    indices.push_back(idx + initial_vtx);
+                    indices.push_back(idx + static_cast<uint32_t>(initial_vtx));
                 });
     }
 }
